+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- Fix the SDMMC driver for ESP-IDF V5.5+
 
 ## [0.45.2] - 2025-01-15